requirement.txt
tensorflow==2.10.0
scikit-learn==1.3.0
numpy==1.26.4
matplotlib==3.7.2
seaborn==0.13.0
pandas==2.1.0
streamlit==1.22.0
librosa==0.10.1
opencv-python==4.12.0.88
pillow==11.3.0
keras==2.10.0
protobuf==3.19.6